Prevented bot from posting when there are no suggestions for the week (#24)
* fixed bugs with longs suggestions and tagging the wrong groups

* added introduction member role

* fixed message error that prevented suggestions

* added in better testing and prevented poll posting when there are no suggestions

* took out of test

+++ musicBot.py
@@ -40,15 +40,29 @@ def runOnce():
         conn = sqlite3.connect('weeklyData.db')
         makeTable = 'CREATE TABLE IF NOT EXISTS weekly (id text PRIMARY KEY, content text NOT NULL); '
         makeTableMovies = 'CREATE TABLE IF NOT EXISTS weeklyMovie (id text PRIMARY KEY, content text NOT NULL); '
+        makeNoWinnerTable = 'CREATE TABLE IF NOT EXISTS noWinner (id text PRIMARY KEY, musicMeeting integer NOT NULL, movieMeeting integer NOT NULL); '
 
         c = conn.cursor()
         c.execute(makeTable)
         c.execute(makeTableMovies)
+        c.execute(makeNoWinnerTable)
+        
         c.execute('SELECT * FROM weekly;')
-
         rows = c.fetchall()
         for row in rows:
             dictionary[str(row[0])] = str(row[1])
+
+        c.execute('SELECT * FROM noWinner;')
+        rows = c.fetchall()
+        if(rows):
+            config.noWinnerMusic = rows[0][1]
+            config.noWinnerMovie = rows[0][2]
+            print("noWinnerMusic = " + str(config.noWinnerMusic) + ", noWinnerMovie = " + str(config.noWinnerMovie))
+        else:
+            conn = sqlite3.connect('weeklyData.db')
+            c = conn.cursor()
+            c.execute('INSERT OR REPLACE INTO noWinner(id, musicMeeting, movieMeeting) VALUES(?,?,?);', ("0", 0, 0))
+            conn.commit()
         
         c.execute('SELECT * FROM weeklyMovie;')
         rows = c.fetchall()
@@ -76,41 +90,57 @@ async def sendPoll():
     if(int(datetime.datetime.now().day) + 1 != getLastMeetingDay()):
         print("Posting poll...")
         channel = bot.get_channel(int(config.announcementChannel))
-        pollString = '/poll "' + config.musicId + ', here is the poll for the album of the week:"'
-        for i, j in dictionary.items():
-            pollString += f' "{i} - {j}"'
-        await channel.send(pollString)
-        async for message in channel.history(limit = 10):
-            if message.author == bot.user:
-                await message.delete()
-                break
-        dictionary.clear()
+        if(dictionary.items()):
+            pollString = '/poll "' + config.musicId + ', here is the poll for the album of the week:"'
+            for i, j in dictionary.items():
+                pollString += f' "{i} - {j}"'
+            await channel.send(pollString)
+            async for message in channel.history(limit = 10):
+                if message.author == bot.user:
+                    await message.delete()
+                    break
+            dictionary.clear()
 
-        # Deletes all of the previous week's suggestions
-        conn = sqlite3.connect('weeklyData.db')
-        c = conn.cursor()
-        c.execute('DELETE FROM weekly;')
-        conn.commit()
+            # Deletes all of the previous week's suggestions
+            conn = sqlite3.connect('weeklyData.db')
+            c = conn.cursor()
+            c.execute('DELETE FROM weekly;')
+            conn.commit()
+        else:
+            config.noWinnerMusic = 1
+            conn = sqlite3.connect('weeklyData.db')
+            c = conn.cursor()
+            c.execute('UPDATE noWinner SET musicMeeting=1')
+            conn.commit()
+            print("No suggestions this week")
 
 # Sends out a poll so people can vote on the movie of the week
 async def sendPollMovie():
     print("Posting poll...")
     channel = bot.get_channel(int(config.announcementChannelMovie))
-    pollString = '/poll "' + config.movieId + ', here is the poll for the movie of the week:"'
-    for i, j in dictionaryMovie.items():
-        pollString += f' "{i} - {j}"'
-    await channel.send(pollString)
-    #async for message in channel.history(limit = 10):
-    #    if message.author == bot.user:
-    #        await message.delete()
-    #        break
-    dictionaryMovie.clear()
+    if(dictionaryMovie.items()):
+        pollString = '/poll "' + config.movieId + ', here is the poll for the movie of the week:"'
+        for i, j in dictionaryMovie.items():
+            pollString += f' "{i} - {j}"'
+        await channel.send(pollString)
+        #async for message in channel.history(limit = 10):
+        #    if message.author == bot.user:
+        #        await message.delete()
+        #        break
+        dictionaryMovie.clear()
 
-    # Deletes all of the previous week's suggestions
-    conn = sqlite3.connect('weeklyData.db')
-    c = conn.cursor()
-    c.execute('DELETE FROM weeklyMovie;')
-    conn.commit()
+        # Deletes all of the previous week's suggestions
+        conn = sqlite3.connect('weeklyData.db')
+        c = conn.cursor()
+        c.execute('DELETE FROM weeklyMovie;')
+        conn.commit()
+    else:
+        config.noWinnerMovie = 1
+        conn = sqlite3.connect('weeklyData.db')
+        c = conn.cursor()
+        c.execute('UPDATE noWinner SET movieMeeting=1')
+        conn.commit()
+        print("No movie suggestions this week")
 
 # Sends out a reminder for the music meeting 30 minutes before
 async def sendReminder():
@@ -128,7 +158,7 @@ async def sendReminderMovie():
 @bot.command(name='suggest')
 async def suggest(ctx, *, arg):
     print("Received suggestion")
-    if(str(ctx.channel.id) == config.suggChannel or str(ctx.channel.id) == config.testChannel):
+    if(str(ctx.channel.id) == config.suggChannel):
         dictionary[str(ctx.author)] = str(arg[:256])
         conn = sqlite3.connect('weeklyData.db')
         c = conn.cursor()
@@ -136,7 +166,7 @@ async def suggest(ctx, *, arg):
         conn.commit()
         await ctx.message.add_reaction("👍")
 
-    elif(str(ctx.channel.id) == config.suggChannelMovie or str(ctx.channel.id) == config.testChannelMovie):
+    elif(str(ctx.channel.id) == config.suggChannelMovie):
         dictionaryMovie[str(ctx.author)] = str(arg[:256])
         conn = sqlite3.connect('weeklyData.db')
         c = conn.cursor()
@@ -152,7 +182,7 @@ async def suggest(ctx, *, arg):
 async def listSuggestions(ctx):
     print("Listing suggestions")
     listString = ""
-    if(str(ctx.channel.id) == config.suggChannel or str(ctx.channel.id) == config.testChannel):
+    if(str(ctx.channel.id) == config.suggChannel):
         if(not dictionary):
             await ctx.send("No suggestions... yet")
         else:
@@ -160,7 +190,7 @@ async def listSuggestions(ctx):
                 listString += f'{k} - {v}\n'
             await ctx.send(listString)
 
-    elif(str(ctx.channel.id) == config.suggChannelMovie or str(ctx.channel.id) == config.testChannelMovie):
+    elif(str(ctx.channel.id) == config.suggChannelMovie):
         if(not dictionaryMovie):
             await ctx.send("No suggestions... yet")
         else:
@@ -171,31 +201,31 @@ async def listSuggestions(ctx):
 # Backup method that can only be called in the TestBot server that puts the album choice poll up in case it fails for some reason
 @bot.command(name='poll')
 async def poll(ctx):
-    if(str(ctx.channel.id) == config.testChannel):
+    if(config.inTest == 1 and str(ctx.channel.id) == config.suggChannel):
         await sendPoll()
 
 # Backup method that can only be called in the TestBot server that puts the album choice poll up in case it fails for some reason
 @bot.command(name='pollMovie')
 async def pollmovie(ctx):
-    if(str(ctx.channel.id) == config.testChannelMovie):
+    if(config.inTest == 1 and str(ctx.channel.id) == config.suggChannelMovie):
         await sendPollMovie()
 
 # Backup method that can only be called in the TestBot server that chooses the winner
 @bot.command(name='choosethewinner')
 async def choosethewinner(ctx):
-    if(str(ctx.channel.id) == config.testChannel):
+    if(config.inTest == 1 and str(ctx.channel.id) == config.suggChannel):
         await chooseWinner()
 
 # Backup method that can only be called in the TestBot server that chooses the winner
 @bot.command(name='choosethewinnermovie')
 async def choosethewinnermovie(ctx):
-    if(str(ctx.channel.id) == config.testChannelMovie):
+    if(config.inTest == 1 and str(ctx.channel.id) == config.suggChannelMovie):
         await chooseWinnerMovie()
 
 # Deletes your suggestion for the week
 @bot.command(name='delete')
 async def delete(ctx):
-    if(str(ctx.channel.id) == config.suggChannel or str(ctx.channel.id) == config.testChannel):
+    if(str(ctx.channel.id) == config.suggChannel):
         conn = sqlite3.connect('weeklyData.db')
         c = conn.cursor()
         c.execute("DELETE FROM weekly WHERE id = ?;", (str(ctx.author)))
@@ -203,7 +233,7 @@ async def delete(ctx):
         dictionary.pop(str(ctx.author))
         await ctx.message.add_reaction("👍")
 
-    elif(str(ctx.channel.id) == config.suggChannelMovie or str(ctx.channel.id) == config.testChannelMovie):
+    elif(str(ctx.channel.id) == config.suggChannelMovie):
         conn = sqlite3.connect('weeklyData.db')
         c = conn.cursor()
         c.execute("DELETE FROM weeklyMovie WHERE id = ?;", (str(ctx.author), ))
@@ -214,9 +244,59 @@ async def delete(ctx):
 # Chooses a winner from the album poll on the Covid Club server
 async def chooseWinner():
     channel = bot.get_channel(int(config.announcementChannel))
+    if(config.noWinnerMusic == 0):
+        if(int(datetime.datetime.now().day) != getLastMeetingDay()):
+            # Finds the last poll posted in the music announcements channel and gets the winner. Myself and the admins are the only ones that can post in that channel.
+            async for message in channel.history(limit = 10):
+                if message.author.id == 324631108731928587:
+                    max = 0
+                    index = []
+                    i = 0
+                    while i <  len(message.reactions):
+                        if(int(message.reactions[i].count) > max):
+                            max = message.reactions[i].count
+                            index = [i]
+                        elif message.reactions[i].count == max:
+                            index.append(i)
+                        i = i+1
+                    await channel.send(config.musicId + 'And the winning album for the week is:')
 
-    if(int(datetime.datetime.now().day) != getLastMeetingDay()):
-        # Finds the last poll posted in the music announcements channel and gets the winner. Myself and the admins are the only ones that can post in that channel.
+                    # Breaks ties with the random function
+                    await channel.send(message.embeds[0].description.split("\n")[random.choice(index)])
+                    if(int(datetime.datetime.now().day) + 7 == getLastMeetingDay()):
+                        await channel.send("It's singles week! If you have a song you want everyone to hear during the meeting next week, use the suggest command with the name of the song and the artist before then!")
+                    else:
+                        await channel.send("Suggestions are now open for the following week, so make sure to get them in by " + calendar.day_name[config.pollDay] + " at " + hourToPrintStandardTime(config.pollHour, config.pollMinute) + "!")
+                    break
+        else:
+            print("Listing singles week songs...")
+            if(not dictionary):
+                await channel.send("No suggestions :(")
+            else:
+                listString = "Here are the songs for singles week: \n"
+                for k, v in dictionary.items():
+                    listString += f'{k} - {v}\n'
+                await channel.send(listString)
+                dictionary.clear()
+
+                # Deletes all of the previous week's suggestions
+                conn = sqlite3.connect('weeklyData.db')
+                c = conn.cursor()
+                c.execute('DELETE FROM weekly;')
+                conn.commit()
+    else:
+        print("No winner chosen for the music club")
+        conn = sqlite3.connect('weeklyData.db')
+        c = conn.cursor()
+        c.execute('UPDATE noWinner SET musicMeeting=0')
+        conn.commit()
+        config.noWinnerMusic = 0
+
+# Chooses a winner from the album poll on the Covid Club server
+async def chooseWinnerMovie():
+    channel = bot.get_channel(int(config.announcementChannelMovie))
+    if(config.noWinnerMovie == 0):
+        # Finds the last poll posted in the movie announcements channel and gets the winner. Myself and the admins are the only ones that can post in that channel.
         async for message in channel.history(limit = 10):
             if message.author.id == 324631108731928587:
                 max = 0
@@ -229,59 +309,24 @@ async def chooseWinner():
                     elif message.reactions[i].count == max:
                         index.append(i)
                     i = i+1
-                await channel.send(config.musicId + 'And the winning album for the week is:')
+                await channel.send(config.movieId + 'And the winning movie for the week is:')
 
                 # Breaks ties with the random function
                 await channel.send(message.embeds[0].description.split("\n")[random.choice(index)])
-                if(int(datetime.datetime.now().day) + 7 == getLastMeetingDay()):
-                    await channel.send("It's singles week! If you have a song you want everyone to hear during the meeting next week, use the suggest command with the name of the song and the artist before then!")
-                else:
-                    await channel.send("Suggestions are now open for the following week, so make sure to get them in by " + calendar.day_name[config.pollDay] + " at " + hourToPrintStandardTime(config.pollHour, config.pollMinute) + "!")
+                await channel.send("Suggestions are now open for the following week, so make sure to get them in by " + calendar.day_name[config.pollDayMovie] + " at " + hourToPrintStandardTime(config.pollHourMovie, config.pollMinuteMovie) + "!")
                 break
     else:
-        print("Listing singles week songs...")
-        if(not dictionary):
-            await channel.send("No suggestions :(")
-        else:
-            listString = "Here are the songs for singles week: \n"
-            for k, v in dictionary.items():
-                listString += f'{k} - {v}\n'
-            await channel.send(listString)
-            dictionary.clear()
-
-            # Deletes all of the previous week's suggestions
-            conn = sqlite3.connect('weeklyData.db')
-            c = conn.cursor()
-            c.execute('DELETE FROM weekly;')
-            conn.commit()
-
-# Chooses a winner from the album poll on the Covid Club server
-async def chooseWinnerMovie():
-    channel = bot.get_channel(int(config.announcementChannelMovie))
-    # Finds the last poll posted in the movie announcements channel and gets the winner. Myself and the admins are the only ones that can post in that channel.
-    async for message in channel.history(limit = 10):
-        if message.author.id == 324631108731928587:
-            max = 0
-            index = []
-            i = 0
-            while i <  len(message.reactions):
-                if(int(message.reactions[i].count) > max):
-                    max = message.reactions[i].count
-                    index = [i]
-                elif message.reactions[i].count == max:
-                    index.append(i)
-                i = i+1
-            await channel.send(config.movieId + 'And the winning movie for the week is:')
-
-            # Breaks ties with the random function
-            await channel.send(message.embeds[0].description.split("\n")[random.choice(index)])
-            await channel.send("Suggestions are now open for the following week, so make sure to get them in by " + calendar.day_name[config.pollDayMovie] + " at " + hourToPrintStandardTime(config.pollHourMovie, config.pollMinuteMovie) + "!")
-            break
+        print("No winner chosen for the movie club")
+        print("No winner chosen for the music club")
+        conn = sqlite3.connect('weeklyData.db')
+        c = conn.cursor()
+        c.execute('UPDATE noWinner SET movieMeeting=0')
+        conn.commit()
+        config.noWinnerMovie = 0
 
 @bot.event
 async def on_ready():
     runOnce()
-
 
 #Checks if the user posted in the introductions channel and gives them member permission
 @bot.event
